+++ CHANGELOG.md
@@ -13,12 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
  -->
 
-
 ## [1.0.10] - 2023-11-18
 
 ### Fixed
 
 - issue where `CarouselItemViewerContainer` would re-calculate optimal height for current item when changing viewing modes.
+- bug causing the text for long section names to not be rendered in the correct position in some cases
 
 ### Added